+++ brzo_i2c.h
@@ -50,8 +50,8 @@ void brzo_i2c_setup(uint32_t clock_stretch_time_out_usec);
 #endif
 
 void brzo_i2c_start_transaction(uint8_t slave_address, uint16_t SCL_frequency_KHz);
-void brzo_i2c_write(uint8_t *data, uint8_t no_of_bytes, bool repeated_start);
-void brzo_i2c_read(uint8_t *data, uint8_t nr_of_bytes, bool repeated_start);
+void brzo_i2c_write(uint8_t *data, uint32_t no_of_bytes, bool repeated_start);
+void brzo_i2c_read(uint8_t *data, uint32_t nr_of_bytes, bool repeated_start);
 uint8_t brzo_i2c_end_transaction();
 
 #ifdef __cplusplus
